+++ GameServer/effects/StagEffect.cs
@@ -89,7 +89,7 @@ namespace DOL.GS.Effects
 				m_amount = (int)((living as GamePlayer).CalculateMaxHealth(living.Level, living.GetModified(eProperty.Constitution)) * m_amountPercent);
 			else m_amount = (int)(living.MaxHealth * m_amountPercent);
 
-			living.BaseBuffBonusCategory[(int)eProperty.MaxHealth] += m_amount;
+			living.BaseBuffBonusCategory[eProperty.MaxHealth] += m_amount;
 			living.Health += (int)(living.GetModified(eProperty.MaxHealth) * m_amountPercent);
 			if (living.Health > living.MaxHealth) living.Health = living.MaxHealth;
 
@@ -109,7 +109,7 @@ namespace DOL.GS.Effects
 
 			double m_amountPercent = m_amount / m_owner.GetModified(eProperty.MaxHealth);
 			int playerHealthPercent = m_owner.HealthPercent;
-			m_owner.BaseBuffBonusCategory[(int)eProperty.MaxHealth] -= m_amount;
+			m_owner.BaseBuffBonusCategory[eProperty.MaxHealth] -= m_amount;
 			if (m_owner.IsAlive)
 				m_owner.Health = (int)Math.Max(1, 0.01 * m_owner.MaxHealth * playerHealthPercent);
