+++ GameServer/propertycalc/BountyPointsCalculator.cs
@@ -32,7 +32,7 @@ namespace DOL.GS.PropertyCalc
 		{
 			if (living is GamePlayer)
 			{
-				return Math.Min(10, living.ItemBonus[(int)property]);
+				return Math.Min(10, living.ItemBonus[property]);
 			}
 
 			return 0;
