+++ GameServer/propertycalc/MythicalCoinCalculator.cs
@@ -16,8 +16,6 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
  */
-
-
 using System;
 
 namespace DOL.GS.PropertyCalc
@@ -32,7 +30,7 @@ namespace DOL.GS.PropertyCalc
         {
             if (living is GamePlayer)
             {
-                return living.ItemBonus[(int)property];
+                return living.ItemBonus[property];
             }
             return 0;
         }
