+++ GameServer/propertycalc/SpellLevelCalculator.cs
@@ -39,12 +39,12 @@ namespace DOL.GS.PropertyCalc
 		public override int CalcValue(GameLiving living, eProperty property)
 		{
 			return (int)(
-				+living.BaseBuffBonusCategory[(int)property]
-				+ living.SpecBuffBonusCategory[(int)property]
-				- living.DebuffCategory[(int)property]
-				+ living.BuffBonusCategory4[(int)property]
-				+ living.AbilityBonus[(int)property]
-				+ Math.Min(10, living.ItemBonus[(int)property]));
+				+living.BaseBuffBonusCategory[property]
+				+ living.SpecBuffBonusCategory[property]
+				- living.DebuffCategory[property]
+				+ living.BuffBonusCategory4[property]
+				+ living.AbilityBonus[property]
+				+ Math.Min(10, living.ItemBonus[property]));
 		}
 	}
 }
