+++ GameServer/propertycalc/WaterSpeedCalculator.cs
@@ -39,11 +39,11 @@ namespace DOL.GS.PropertyCalc
 		public override int CalcValue(GameLiving living, eProperty property)
 		{
 			return (int)(
-				+ living.BaseBuffBonusCategory[(int)property]
-				+ living.SpecBuffBonusCategory[(int)property]
-				- living.DebuffCategory[(int)property]
-				+ living.BuffBonusCategory4[(int)property]
-                + living.ItemBonus[(int)property]);
+				+ living.BaseBuffBonusCategory[property]
+				+ living.SpecBuffBonusCategory[property]
+				- living.DebuffCategory[property]
+				+ living.BuffBonusCategory4[property]
+                + living.ItemBonus[property]);
 		}
 	}
 }
